requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
sentence-transformers
faiss-cpu
pandas
numpy
sqlalchemy

